Replace 'uint' with 'uint256'

+++ LUSDChickenBonds/src/ChickenBondManager.sol
@@ -31,7 +31,7 @@ contract ChickenBondManager is Ownable {
     }
 
     uint256 public totalPendingLUSD;
-    mapping (uint => BondData) public idToBondData;
+    mapping (uint256 => BondData) public idToBondData;
 
     uint256 constant MAX_UINT256 = type(uint256).max;
     uint256 constant SECONDS_IN_ONE_MONTH = 2592000;
@@ -91,7 +91,7 @@ contract ChickenBondManager is Ownable {
     * TODO: Decide if we want on-chain functionality for returning a list of a given bonder's NFTs. Increases minting gas cost.
     */
 
-    function chickenOut(uint _bondID) external {
+    function chickenOut(uint256 _bondID) external {
         _requireCallerOwnsBond(_bondID);
 
           /* In practice this assert should mostly hold. However, there could be edge cases where it doesn't: 
@@ -104,12 +104,12 @@ contract ChickenBondManager is Ownable {
         */
         assert(getLUSDInYearn() - getAcquiredLUSDInYearn() >= totalPendingLUSD);
 
-        uint bondedLUSD = idToBondData[_bondID].lusdAmount;
+        uint256 bondedLUSD = idToBondData[_bondID].lusdAmount;
         
         delete idToBondData[_bondID];
         totalPendingLUSD -= bondedLUSD;
 
-        uint yTokensToBurn = yearnLUSDVault.calcYTokenToToken(bondedLUSD);
+        uint256 yTokensToBurn = yearnLUSDVault.calcYTokenToToken(bondedLUSD);
         yearnLUSDVault.withdraw(yTokensToBurn);
 
         // Send bonded LUSD back to caller and burn their bond NFT
@@ -117,21 +117,21 @@ contract ChickenBondManager is Ownable {
         bondNFT.burn(_bondID);
     }
 
-    function chickenIn(uint _bondID) external {
+    function chickenIn(uint256 _bondID) external {
         _requireCallerOwnsBond(_bondID);
 
         BondData memory bond = idToBondData[_bondID];
         
-        uint backingRatio = calcSystemBackingRatio();
-        uint accruedSLUSD = _calcAccruedSLUSD(bond, backingRatio);
+        uint256 backingRatio = calcSystemBackingRatio();
+        uint256 accruedSLUSD = _calcAccruedSLUSD(bond, backingRatio);
     
         delete idToBondData[_bondID];
         totalPendingLUSD -= bond.lusdAmount;
 
         /* Get LUSD amounts to acquire and refund. Acquire LUSD in proportion to the system's current backing ratio, 
         * in order to maintain it. */
-        uint lusdToAcquire = accruedSLUSD * backingRatio / 1e18;
-        uint lusdToRefund = bond.lusdAmount - lusdToAcquire;
+        uint256 lusdToAcquire = accruedSLUSD * backingRatio / 1e18;
+        uint256 lusdToRefund = bond.lusdAmount - lusdToAcquire;
         
         // Pull the refund from Yearn LUSD vault
         yearnLUSDVault.withdraw(lusdToRefund);
@@ -142,32 +142,32 @@ contract ChickenBondManager is Ownable {
         bondNFT.burn(_bondID);
     }
 
-    function redeem(uint _sLUSDToRedeem) external {
+    function redeem(uint256 _sLUSDToRedeem) external {
         /* TODO: determine whether we should simply leave the fee in the acquired bucket, or add it to a permanent bucket.
         Current approach leaves redemption fees in the acquired bucket. */
-        uint fractionOfSLUSDToRedeem = _sLUSDToRedeem * 1e18 / sLUSDToken.totalSupply();
+        uint256 fractionOfSLUSDToRedeem = _sLUSDToRedeem * 1e18 / sLUSDToken.totalSupply();
     
         // Calculate redemption fraction to withdraw, given that we leave the fee inside the system
-        uint fractionOfAcquiredLUSDToWithdraw = fractionOfSLUSDToRedeem * (1e18 - calcRedemptionFeePercentage()) / 1e18;
+        uint256 fractionOfAcquiredLUSDToWithdraw = fractionOfSLUSDToRedeem * (1e18 - calcRedemptionFeePercentage()) / 1e18;
 
         // Get the LUSD to withdraw from Yearn, and the corresponding yTokens
-        uint lusdToWithdrawFromYearn = getAcquiredLUSDInYearn() * fractionOfAcquiredLUSDToWithdraw / 1e18;
-        uint yTokensToWithdrawFromLUSDVault = yearnLUSDVault.calcTokenToYToken(lusdToWithdrawFromYearn);
+        uint256 lusdToWithdrawFromYearn = getAcquiredLUSDInYearn() * fractionOfAcquiredLUSDToWithdraw / 1e18;
+        uint256 yTokensToWithdrawFromLUSDVault = yearnLUSDVault.calcTokenToYToken(lusdToWithdrawFromYearn);
         
         // Since 100% of the Curve liquidity is "acquired", just get the yTokens directly
-        uint yTokensToWithdrawFromCurveVault = yearnCurveVault.balanceOf(address(this)) * fractionOfAcquiredLUSDToWithdraw / 1e18;
+        uint256 yTokensToWithdrawFromCurveVault = yearnCurveVault.balanceOf(address(this)) * fractionOfAcquiredLUSDToWithdraw / 1e18;
 
         // The LUSD and LUSD3CRV deltas from SP/Curve withdrawals are the amounts to send to the redeemer
-        uint lusdBalanceBefore = lusdToken.balanceOf(address(this));
-        uint LUSD3CRVBalanceBefore = curvePool.balanceOf(address(this));
+        uint256 lusdBalanceBefore = lusdToken.balanceOf(address(this));
+        uint256 LUSD3CRVBalanceBefore = curvePool.balanceOf(address(this));
 
         yearnLUSDVault.withdraw(yTokensToWithdrawFromLUSDVault); // obtain LUSD from Yearn
         yearnCurveVault.withdraw(yTokensToWithdrawFromCurveVault); // obtain LUSD3CRV from Yearn
 
-        uint LUSD3CRVDelta = curvePool.balanceOf(address(this)) - LUSD3CRVBalanceBefore;
+        uint256 LUSD3CRVDelta = curvePool.balanceOf(address(this)) - LUSD3CRVBalanceBefore;
         curvePool.remove_liquidity(LUSD3CRVDelta); // obtain LUSD from Curve
 
-        uint lusdBalanceDelta = lusdToken.balanceOf(address(this)) - lusdBalanceBefore;
+        uint256 lusdBalanceDelta = lusdToken.balanceOf(address(this)) - lusdBalanceBefore;
 
         // Burn the redeemed sLUSD
         sLUSDToken.burn(msg.sender, _sLUSDToRedeem);
@@ -178,15 +178,15 @@ contract ChickenBondManager is Ownable {
 
     function shiftLUSDFromSPToCurve() external {
         // Calculate the LUSD to pull from the Yearn LUSD vault
-        uint lusdToShift = _calcLUSDToShiftToCurve();
+        uint256 lusdToShift = _calcLUSDToShiftToCurve();
         _requireNonZeroLUSDToShift(lusdToShift);
 
-        uint yTokensToBurn = yearnLUSDVault.calcTokenToYToken(lusdToShift);
+        uint256 yTokensToBurn = yearnLUSDVault.calcTokenToYToken(lusdToShift);
 
         // Convert yTokens to LUSD
-        uint lusdBalanceBefore = lusdToken.balanceOf(address(this));
+        uint256 lusdBalanceBefore = lusdToken.balanceOf(address(this));
         yearnLUSDVault.withdraw(yTokensToBurn);
-        uint lusdBalanceDelta = lusdToken.balanceOf(address(this)) - lusdBalanceBefore;
+        uint256 lusdBalanceDelta = lusdToken.balanceOf(address(this)) - lusdBalanceBefore;
 
         /* In principle this assert should hold. In practice, there may be a slight discrepancy depending on Yearn calculations
         * (in which case, we should decide on the acceptable margin of tolerance).
@@ -194,9 +194,9 @@ contract ChickenBondManager is Ownable {
         assert(lusdBalanceDelta == lusdToShift);
 
         // Deposit the received LUSD to Curve in return for LUSD3CRV-f tokens
-        uint LUSD3CRVBalanceBefore = curvePool.balanceOf(address(this));
+        uint256 LUSD3CRVBalanceBefore = curvePool.balanceOf(address(this));
         curvePool.add_liquidity(lusdBalanceDelta);
-        uint LUSD3CRVBalanceDelta = curvePool.balanceOf(address(this)) - LUSD3CRVBalanceBefore;
+        uint256 LUSD3CRVBalanceDelta = curvePool.balanceOf(address(this)) - LUSD3CRVBalanceBefore;
 
         // Deposit the received LUSD3CRV-f to Yearn Curve vault
         yearnCurveVault.deposit(LUSD3CRVBalanceDelta);
@@ -204,19 +204,19 @@ contract ChickenBondManager is Ownable {
 
    function shiftLUSDFromCurveToSP() external {
        // Calculate LUSD to pull from Curve
-        uint lusdToShift = _calcLUSDToShiftToSP();
+        uint256 lusdToShift = _calcLUSDToShiftToSP();
         _requireNonZeroLUSDToShift(lusdToShift);
         
         //Calculate LUSD3CRV-f needed to withdraw LUSD from Curve
-        uint LUSD3CRVfToBurn = curvePool.calcLUSDToLUSD3CRV(lusdToShift);
+        uint256 LUSD3CRVfToBurn = curvePool.calcLUSDToLUSD3CRV(lusdToShift);
 
         //Calculate yTokens to swap for LUSD3CRV-f 
-        uint yTokensToBurn = yearnCurveVault.calcTokenToYToken(LUSD3CRVfToBurn);
+        uint256 yTokensToBurn = yearnCurveVault.calcTokenToYToken(LUSD3CRVfToBurn);
 
         // Convert yTokens to LUSD3CRV-f
-        uint LUSD3CRVBalanceBefore = curvePool.balanceOf(address(this));
+        uint256 LUSD3CRVBalanceBefore = curvePool.balanceOf(address(this));
         yearnCurveVault.withdraw(yTokensToBurn);
-        uint LUSD3CRVBalanceDelta = curvePool.balanceOf(address(this)) - LUSD3CRVBalanceBefore;
+        uint256 LUSD3CRVBalanceDelta = curvePool.balanceOf(address(this)) - LUSD3CRVBalanceBefore;
 
         /* In principle this assert should hold. In practice, there may be a slight discrepancy depending on Yearn calculations
         * (in which case, we should decide on the acceptable margin of tolerance).
@@ -224,9 +224,9 @@ contract ChickenBondManager is Ownable {
         assert(LUSD3CRVBalanceDelta == LUSD3CRVfToBurn);
 
         // Withdraw LUSD from Curve
-        uint lusdBalanceBefore = lusdToken.balanceOf(address(this));
+        uint256 lusdBalanceBefore = lusdToken.balanceOf(address(this));
         curvePool.remove_liquidity(LUSD3CRVBalanceDelta);
-        uint lusdBalanceDelta = lusdToken.balanceOf(address(this)) - lusdBalanceBefore;
+        uint256 lusdBalanceDelta = lusdToken.balanceOf(address(this)) - lusdBalanceBefore;
 
         // Should hold in principle. Depends on Curve calculations
         assert(lusdBalanceDelta == lusdToShift);
@@ -261,7 +261,7 @@ contract ChickenBondManager is Ownable {
     }
 
     // External getter for calculating accrued LUSD based on bond ID
-    function calcAccruedSLUSD(uint _bondID) external view returns (uint256) {
+    function calcAccruedSLUSD(uint256 _bondID) external view returns (uint256) {
         BondData memory bond = idToBondData[_bondID];
 
         return _calcAccruedSLUSD(bond, calcSystemBackingRatio());
@@ -278,8 +278,8 @@ contract ChickenBondManager is Ownable {
         * results in an accrued sLUSD equal to 50% of the cap after one month.
         *
         * TODO: replace with final sLUSD accrual formula. */
-        uint bondDuration = (block.timestamp - _bond.startTime);
-        uint accruedSLUSD = bondSLUSDCap * bondDuration / (bondDuration + SECONDS_IN_ONE_MONTH);
+        uint256 bondDuration = (block.timestamp - _bond.startTime);
+        uint256 accruedSLUSD = bondSLUSDCap * bondDuration / (bondDuration + SECONDS_IN_ONE_MONTH);
         assert(accruedSLUSD < bondSLUSDCap);
 
         return accruedSLUSD;
@@ -305,14 +305,14 @@ contract ChickenBondManager is Ownable {
     }
 
     function getLUSDInYearn() public view returns (uint256) {
-        uint yTokenBalanceLUSD = yearnLUSDVault.balanceOf(address(this));
-        uint lusdInYearn = yearnLUSDVault.calcYTokenToToken(yTokenBalanceLUSD);
+        uint256 yTokenBalanceLUSD = yearnLUSDVault.balanceOf(address(this));
+        uint256 lusdInYearn = yearnLUSDVault.calcYTokenToToken(yTokenBalanceLUSD);
 
         return lusdInYearn;
     }
 
     function getAcquiredLUSDInYearn() public view returns (uint256) {
-        uint acquiredLUSDInYearn = getLUSDInYearn() - totalPendingLUSD;
+        uint256 acquiredLUSDInYearn = getLUSDInYearn() - totalPendingLUSD;
 
         assert(acquiredLUSDInYearn >= 0);
 
@@ -320,17 +320,17 @@ contract ChickenBondManager is Ownable {
     }
 
     function getAcquiredLUSDInCurve() public view returns (uint256) {
-        uint yTokenBalanceLUSD3CRV = yearnCurveVault.balanceOf(address(this));
-        uint lusd3CRVInYearn = yearnCurveVault.calcYTokenToToken(yTokenBalanceLUSD3CRV);
+        uint256 yTokenBalanceLUSD3CRV = yearnCurveVault.balanceOf(address(this));
+        uint256 lusd3CRVInYearn = yearnCurveVault.calcYTokenToToken(yTokenBalanceLUSD3CRV);
         
-        uint lusdInCurve = curvePool.calcLUSD3CRVToLUSD(lusd3CRVInYearn);
+        uint256 lusdInCurve = curvePool.calcLUSD3CRVToLUSD(lusd3CRVInYearn);
 
         return lusdInCurve;
     }
 
     function calcSystemBackingRatio() public view returns (uint256) {
-        uint totalSLUSDSupply = sLUSDToken.totalSupply();
-        uint totalAcquiredLUSD = getTotalAcquiredLUSD();
+        uint256 totalSLUSDSupply = sLUSDToken.totalSupply();
+        uint256 totalAcquiredLUSD = getTotalAcquiredLUSD();
 
         /* TODO: Determine how to define the backing ratio when there is 0 sLUSD and 0 totalAcquiredLUSD,
         * i.e. before the first chickenIn. For now, return a backing ratio of 1. Note: Both quantities would be 0
@@ -344,14 +344,14 @@ contract ChickenBondManager is Ownable {
 
     // External getter for calculating the bond sLUSD cap based on bond ID
     function calcBondSLUSDCap(uint256 _bondID) external view returns (uint256) {
-        uint backingRatio = calcSystemBackingRatio();
+        uint256 backingRatio = calcSystemBackingRatio();
         BondData memory bond = idToBondData[_bondID];
 
         return _calcBondSLUSDCap(bond.lusdAmount, backingRatio);
     }
 
     // Internal getter for calculating the bond sLUSD cap based on bonded amount and backing ratio
-    function _calcBondSLUSDCap(uint _bondedAmount, uint _backingRatio) internal view returns (uint256) {
+    function _calcBondSLUSDCap(uint256 _bondedAmount, uint256 _backingRatio) internal view returns (uint256) {
         // TODO: potentially refactor this -  i.e. have a (1 / backingRatio) function for more precision
         return _bondedAmount * 1e18 / _backingRatio;
     }
